introduce AbstractRenderService to reuse code
