+++ webapp/src/main/resources/static/ts/src/service/my-artists-render-service.ts
@@ -6,62 +6,52 @@ import { AlertService } from "./alert-service";
 import { Pagination } from "../model/pagination.model";
 import { LoadingIndicatorService } from "./loading-indicator-service";
 import { PaginationComponent } from "../components/pagination/pagination-component";
+import { AbstractRenderService } from "./abstract-render-service";
 
-export class MyArtistsRenderService {
+export class MyArtistsRenderService extends AbstractRenderService<MyArtistsResponse> {
 
     private readonly MAX_CARDS_PER_ROW = 4;
-    private readonly HOST_ELEMENT_ID = "artists-container";
 
     private readonly followArtistService: FollowArtistService;
-    private readonly alertService: AlertService;
-    private readonly loadingIndicatorService: LoadingIndicatorService;
     private readonly paginationComponent: PaginationComponent;
     private readonly artistTemplateElement: HTMLTemplateElement;
-    private readonly hostElement: HTMLDivElement;
     private rowElement?: HTMLDivElement;
 
     constructor(followArtistService: FollowArtistService, alertService: AlertService, loadingIndicatorService: LoadingIndicatorService) {
+        super(alertService, loadingIndicatorService);
         this.followArtistService = followArtistService;
-        this.alertService = alertService;
-        this.loadingIndicatorService = loadingIndicatorService;
         this.paginationComponent = new PaginationComponent();
         this.artistTemplateElement = document.getElementById("artist-card")! as HTMLTemplateElement;
-        this.hostElement = document.getElementById(this.HOST_ELEMENT_ID)! as HTMLDivElement;
     }
 
-    public renderResults(data: Promise<MyArtistsResponse>): void {
-        this.loadingIndicatorService.showLoadingIndicator(this.HOST_ELEMENT_ID);
-        data.then((response) => {
-            if (response.myArtists.length === 0 && response.pagination.currentPage === 1) {
-                const message = "Currently you do not follow any artist. Start a search for your favorite artists right now.";
-                const infoMessage = this.alertService.renderInfoAlert(message, false);
-                this.hostElement.insertAdjacentElement("afterbegin", infoMessage);
-            }
-            else if (response.pagination.totalPages < response.pagination.currentPage) {
-                // If you follow for example 21 bands and you unfollow one band on page 1 and then go to page 2 there are no bands to display.
-                // In this case you will be forwarded to the last page.
-                window.location.replace(`?page=${response.pagination.totalPages}`);
-            }
-            else {
-                let currentCardNo = 1;
-                response.myArtists.forEach((artist => {
-                    const artistDivElement = this.renderArtistCard(artist);
-                    this.attachArtistCard(artistDivElement, currentCardNo);
-                    currentCardNo++;
-                }));
+    protected getHostElementId(): string {
+        return "artists-container";
+    }
 
-                const pagination = response.pagination;
-                if (pagination.totalPages > 1) {
-                    this.attachPagination(pagination);
-                }
-            }
-        }).catch(() => {
-            const message = "Damn it! A satanic error has occurred. Please try again later.";
-            const infoMessage = this.alertService.renderErorAlert(message, false);
+    protected onRendering(response: MyArtistsResponse): void {
+        if (response.myArtists.length === 0 && response.pagination.currentPage === 1) {
+            const message = "Currently you do not follow any artist. Start a search for your favorite artists right now.";
+            const infoMessage = this.alertService.renderInfoAlert(message, false);
             this.hostElement.insertAdjacentElement("afterbegin", infoMessage);
-        }).finally(() => {
-            this.loadingIndicatorService.hideLoadingIndicator(this.HOST_ELEMENT_ID);
-        });
+        }
+        else if (response.pagination.totalPages < response.pagination.currentPage) {
+            // If you follow for example 21 bands and you unfollow one band on page 1 and then go to page 2 there are no bands to display.
+            // In this case you will be forwarded to the last page.
+            window.location.replace(`?page=${response.pagination.totalPages}`);
+        }
+        else {
+            let currentCardNo = 1;
+            response.myArtists.forEach((artist => {
+                const artistDivElement = this.renderArtistCard(artist);
+                this.attachArtistCard(artistDivElement, currentCardNo);
+                currentCardNo++;
+            }));
+
+            const pagination = response.pagination;
+            if (pagination.totalPages > 1) {
+                this.attachPagination(pagination);
+            }
+        }
     }
 
     private renderArtistCard(artist: Artist): HTMLDivElement {
